Move functionality to new register_validator.go file
